data_gen/requirements.txt
annotated-types==0.7.0
apache-beam==2.56.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
Faker==25.2.0
fastavro==1.9.4
fasteners==0.19
google-api-core==2.19.0
google-apitools==0.5.31
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.54.1
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.23.1
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-dlp==3.18.0
google-cloud-language==2.13.3
google-cloud-pubsub==2.21.1
google-cloud-pubsublite==1.10.0
google-cloud-recommendations-ai==0.10.10
google-cloud-resource-manager==1.12.3
google-cloud-spanner==3.47.0
google-cloud-storage==2.17.0
google-cloud-videointelligence==2.13.3
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.7.1
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.64.0
grpcio-status==1.62.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
jsonpickle==3.0.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==1.26.4
oauth2client==4.1.3
objsize==0.7.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
pydot==1.4.2
pymongo==4.7.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rpds-py==0.18.1
rsa==4.9
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
zstandard==0.22.0
